Time and RTC changes
* Add RTC::micros mutator - This will also allow other clock-correction
techniques.  Perhaps the current correction technique should be made
external to RTC: it only makes `seconds` as accurate as the crystal, it
imposes the overhead on a majority of apps that don't care about
`seconds` accuracy, and it may make a better clock-correction technique
more difficult.
* Add packed to time_t - Allows time_t to be a member of a packed
struct, like gps_fix_t.
* Add `const` to `IOStream & operator<<( time_t &)` - Allows it to work
on const time_t instead of implicitly casting to long clock_t.  Perhaps
this should be considered for other `IOStream & operator <<`.
*Minor change in OVF IRQ - Set `s_ticks` just once.