Fix make size checking
If a build is successful and required flash size was ok it creates a
file with a suffix of ‘.sizeok’. This is used to allow uploads later.
If the application grows and subsequent build exceeds the size limit
the ‘.sizeok’ file was NOT removed resulting in an upload that
overwrites the bootloader.  ‘.sizeok’ was only being removed via
‘clean’.  The fix now removes ‘.sizeok’ correctly.