Better high-rate Event handling
Event queue is very short and can easily cause high-rate events to be lost. This change adds 1 byte to each queue slot to maintain a counter of the number of "push" events that were recorded for that slot (ie matching events). The next slot will be taken if a push would cause a counter overflow or it is a different event.

The main negative to this change is that any class used in a queue must implement the "==" operator. A little overhead is added to enqueue/dequeue.