Handle PySerial changes and support advanced adapters.
Newer versions of PySerial have removed setBaudrate and replaced with the attribute baudrate. To maintain backwards compatibility, this change tries the old method and if it fails it then uses the new method.

In addition, enhanced behavior for advanced USB to serial adapters has been implemented. The details are explained in comments in the file. This change should not have any affect on standard USB to serial adapters.